tests/wpt/css-tests/cssom-view-1_dev/html/window-screen-width-immutable.htm

<!DOCTYPE html>
<html><head>
    <title>CSSOM View - 4.2 - screen.width immutability</title>
    <link href="mailto:neils.christoffersen@gmail.com" rel="author" title="Neils Christoffersen">
    <link href="http://www.w3.org/TR/cssom-view/#the-screen-interface" rel="help">
    <meta content="dom" name="flags">
    <meta content="screen.width is immutable" name="assert">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body>
    <div id="myDiv"></div>
    <div id="log"></div>
    <script>
		var originalVal = window.screen.width;

		// try to set window.screen.width
		window.screen.width = 0;

		// verify window.screen.width didn't change
        test(function() {
			assert_equals(window.screen.width, originalVal,	"window.screen.width should be immutable")
		}, "immutability test");
    </script>


</body></html>
